Junit structure changes and update
